+++ CHANGELOG.md
@@ -9,7 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-* None yet!
+### Updated
+
+* Better exception for smart enum errors
 
 ## [0.25.0] - 12/28/2023